made some modifications
